Enable optimized Parquet writer by default in Delta Lake
