--- plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeCreateTableStatisticsOptimizedWriter.java
@@ -1,28 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.trino.plugin.deltalake;
-
-import com.google.common.collect.ImmutableMap;
-
-import java.util.Map;
-
-public class TestDeltaLakeCreateTableStatisticsOptimizedWriter
-        extends AbstractTestDeltaLakeCreateTableStatistics
-{
-    @Override
-    Map<String, String> additionalProperties()
-    {
-        return ImmutableMap.of("parquet.experimental-optimized-writer.enabled", "true");
-    }
-}
